+++ package.json
@@ -15,6 +15,7 @@
     "clsx": "^2.1.1",
     "cobe": "^0.6.3",
     "framer-motion": "^12.4.7",
+    "framer-motion-3d": "^12.4.7",
     "gsap": "^3.12.7",
     "jq": "^1.7.2",
     "lenis": "^1.1.20",
