+++ src/app/page.js
@@ -17,7 +17,6 @@ import ContactBlock from "./components/Home/contact/contactBlock";
 import Crosslink from "./components/Home/CrossLink/crossLink";
 import LocationSlider from "./components/Home/location_slider/locationSlider";
 import WhatsAppIcon from "./components/Whatsapp/whatsapp";
-import Earth from './components/About/Globe/globe'
 
 
 
@@ -50,7 +49,6 @@ export default function Home() {
           <ContactBlock />
           <WhatsAppIcon/>
           <Footer />
-          <Earth/>
         </>
       )}
     </>
