+++ src/assets/global.css
@@ -58,8 +58,8 @@
 /* Define font2 */
 @font-face {
     font-family: 'Exo_Medium_Italic';
-    src: url('/fonts/Exo_Medium_Italic.woff2') format('woff2'),
-    url('/fonts/Exo_Medium_Italic.woff2') format('woff');
+    src: url('/fonts/Exo_Medium-Italic.woff2') format('woff2'),
+    url('/fonts/Exo_Medium-Italic.woff2') format('woff');
     font-weight: Italic;
     font-style: italic;
 }
@@ -75,16 +75,16 @@
 /* Define font2 */
 @font-face {
     font-family: 'SchibstedGrotesk_Bold';
-    src: url('/fonts/SchibstedGrotesk_Bold') format('woff2'),
-    url('/fonts/SchibstedGrotesk_Bold') format('woff');
+    src: url('/fonts/SchibstedGrotesk-Bold') format('woff2'),
+    url('/fonts/SchibstedGrotesk-Bold') format('woff');
     font-weight: normal;
     font-style: normal;
 }
 /* Define font1 */
 @font-face {
     font-family: 'SchibstedGrotesk_BoldItalic';
-    src: url('/fonts/SchibstedGrotesk_BoldItalic.woff2') format('woff2'),
-         url('/fonts/SchibstedGrotesk_BoldItalic.woff') format('woff');
+    src: url('/fonts/SchibstedGrotesk-BoldItalic.woff2') format('woff2'),
+         url('/fonts/SchibstedGrotesk-BoldItalic.woff') format('woff');
     font-weight: bold;
     font-style: normal;
   }
@@ -92,8 +92,8 @@
 /* Define font2 */
 @font-face {
     font-family: 'SchibstedGrotesk_semibold';
-    src: url('/fonts/SchibstedGrotesk_semibold.woff2') format('woff2'),
-    url('/fonts/SchibstedGrotesk_semiboldß.woff') format('woff');
+    src: url('/fonts/SchibstedGrotesk-semibold.woff2') format('woff2'),
+    url('/fonts/SchibstedGrotesk-semiboldß.woff') format('woff');
     font-weight: normal;
     font-style: normal;
 }
@@ -101,8 +101,8 @@
 /* Define font2 */
 @font-face {
     font-family: 'SchibstedGrotesk_Medium';
-    src: url('/fonts/SchibstedGrotesk_Medium') format('woff2'),
-    url('/fonts/SchibstedGrotesk_Medium.woff') format('woff');
+    src: url('/fonts/SchibstedGrotesk-Medium.woff2') format('woff2'),
+    url('/fonts/SchibstedGrotesk-Medium.woff') format('woff');
     font-weight: normal;
     font-style: normal;
 }
